- Abandon du support pour nouvelles installations sur Debian 4.  - Legere amelioration du script d'install.  - Bug fix (associer a une nouvelle version si pas d'ECTS).

+++ config/upgrade.sh
@@ -12,7 +12,10 @@ source utils.sh
 check_uid_root $0
 
 apt-get update
-apt-get -y install curl # now necessary
+
+if [ ! -e /usr/bin/curl ]; then
+  apt-get -y install curl # now necessary
+fi
 
 echo "Stopping ScoDoc..."
 /etc/init.d/scodoc stop
